requirements.txt
uTidylib==0.6
PyYAML==5.3
requests
